requirements.txt
# ── Core ──────────────────────────────────────────────
pydantic>=2.6,<3.0
python-dotenv>=1.0

# ── ML / Explainability ─────────────────────────────
scikit-learn>=1.5
numpy>=2.0
shap>=0.45
joblib>=1.3
pandas>=2.0

# ── LLM ──────────────────────────────────────────────
groq>=0.5.0

# ── Networking / Link Checking ───────────────────────
httpx>=0.27,<1.0
tldextract>=5.1
beautifulsoup4>=4.12

# ── Frontend (reserved for teammate) ─────────────────
streamlit>=1.35,<2.0

# ── Dev / Testing ────────────────────────────────────
pytest>=8.0
